A bunch of small fixes.
Added assertions to MVC to check for setting null components, and fixed
a go back bug that was pointing to a nonexistent component.
Added pizza place logos so that the providers carousel looked more
realistic.
Removed Utils.lua, as it's a relic of Darwin's Factory.
Fixed issues where subviews were added both to the screen and to
a parent view.


Former-commit-id: 608586a6b0ed16d4c5ee26b079fc04ab44ca7c2a