[engine] Added 'trickplay.config' and 'trickplay.system'.
Config is a table of all the configuration parameters to trickplay. This
table will be empty in production builds so it doesn't reveal paths to the app.
System is a table with basic information about the system that TrickPlay is
running on, such as its name, version, language, country and uuid.


Former-commit-id: f3b54fc8a148df2f53f05f126d2bd5c9b0c75dea
Former-commit-id: 2fc905acd0e960c9732915e6198b7ed4369aa502