[engine] Added 'trickplay.config' and 'trickplay.system'.
Config is a table of all the configuration parameters to trickplay. This
table will be empty in production builds so it doesn't reveal paths to the app.
System is a table with basic information about the system that TrickPlay is
running on, such as its name, version, language, country and uuid.