[iphone] Fixed a bug where images were being squished on the y-axis.
Previously, when images were written to the phone within the 320x435 view (adjusted because of the top navigation bar) the images were scaled in the y-axis and placed in the 320x435 view as if the dimensions were 320x480. I.E. an image of y = 60 and hieght = 30 would show up on the screen at y = 60*435/480 and height = 30*435/480, yet still within the scope of the 320x435 view (below the navigation bar).


Former-commit-id: 858e6aefba276feb686b8b1a0914414d90019cac
Former-commit-id: 378c883ead41ea51aad543ab46dc9ce0647dcfb0