[rex-iphone] App now capable of both sending and receiving messages from Trickplay.
The iPhone app can connect to a mDNS server and if that server is a Trickplay server it can communicate using Trickplay's line protocol for controller interaction. However, even though Trickplay will detect the iPod controller and send messages, currently the messages are simply parsed and printed to the iPod's console. Commands are not yet interpretted.


Former-commit-id: d97e1473cadf61e555248f754e19cd653226a665
Former-commit-id: d4646cc7dae2348d8453f4112248c091084b246d