[engine] Significant changes to controller API. Added pointer events, removed clicks.
The simple 'click' API is now gone, in favor of a better 'pointer' API for controllers.
The default controller added in the emulator now uses this API for mouse
events.

The touch API adds a finger parameter as well.

This introduces several breaking changes.

For the OEM interface:

        Removed TP_CONTROLLER_HAS_CLICKS
        Removed TP_CONTROLLER_COMMAND_START_CLICKS
        Removed TP_CONTROLLER_COMMAND_STOP_CLICKS
        Removed tp_controller_click
        Changed parameters for tp_controller_touch_down
        Changed parameters for tp_controller_touch_move
        Changed parameters for tp_controller_touch_up

For the Lua interface:

    controller

        Removed on_click
        Removed start_clicks
        Removed stop_clicks
        Changed parameteres for on_touch_down
        Changed parameteres for on_touch_move
        Changed parameteres for on_touch_up