[rex-iphone] Made some pleasent aesthetic changes. Got Updating and Clearing Background/Foreground Images working, albiet slow. Got Touch events to send over to Trickplay form the iPhone.
Trickplay Lua Developer now has full control of iphone in setting images and background to the UI and clearing the UI. However, images are pull currently not cached and must be re-pulled every time they are added to the iphone's view, even if the images were previously there. Additionally, touch events received on the iphone's view now are sent to Trickplay for receipt allowing the Lua Developer to respond to the touch events. Also, some aesthetic changes were made, for instance, it automatically brings the iphone user to a loading screen after selecting a service rather than hanging until a connection is established with the service.


Former-commit-id: a6a30a1918b26cb95d05e90e4f1efa7d3445a90e
Former-commit-id: dfb5bd86777aef55a659a2fe88977e7b9d92b7f5