[rex-iphone] Multithreaded image loading. Touch and Accelerometer controls. Improved speed. Aesthetic improvements. Works with new Trickplay
Sending images to the iphone via Trickplay functions asynchronously. A spinner icon will appear in place of the image until the image loads to the screen and all other functionality works in the meantime. However, updating the background is NOT asynchronous, this is for aesthetic reasons. Updated the nib files and some code so that the objects on screen are positioned correctly on the iPad as well as the iPhone/Pod. Image loading asynchronously uses multiple threads, this has seemed to increase speed though no definitive tests were performed to prove this. A couple of changes were made in order to follow Trickplay 1.16.2's protocol.


Former-commit-id: 451e8ccdc660077102eb17d3f18fe1a6df673ceb
Former-commit-id: 1107c56ce66a820c61fd2e08cbf4afa358a07b6c