[engine] Controllers added and removed in main thread.
Before, controllers were added and removed in the calling thread
and an event was posted to fire the callbacks. This means that a
controller can be in the connected list (and an app can see it) before
on_connected is invoked. Now, adding and removing both post an event
and the list is populated when the event is processed (next time
around the event loop)