[engine] Added app splash image support. Changed app notification.
Apps are now loaded after we enter the main loop. If an app has a file
called 'default.png/jpg' in its root directory, the engine loads and
displays this image before it loads the app itself. This gets the app
to the screen faster.

Notifications for app loaded and closing are only sent to the app - not
the outside world. They were not very meaningful there since you didn't
know what app was affected.