[rex-iphone] Fixed bug where touch up events were sending touch move messages.
Previously if the user touched the screen with a finger, upon lifting the finger, a 'TM' or touch moved event would be sent to Trickplay. This has been changed to the desired effect of the iOS app sending 'TU' or a touch up event to Trickplay upon removing ones finger from the iPod/Pad screen.