[rex-iphone] Fixed bug where touch up events were sending touch move messages.
Previously if the user touched the screen with a finger, upon lifting the finger, a 'TM' or touch moved event would be sent to Trickplay. This has been changed to the desired effect of the iOS app sending 'TU' or a touch up event to Trickplay upon removing ones finger from the iPod/Pad screen.


Former-commit-id: 5730ff1f9856677ebdb91557d9eb6f55ab0df9f7
Former-commit-id: 1d097163fb37b92eee822d4ec558eb7b6b3cf63f