[engine] Added configuration key TP_FIRST_APP_EXITS.
When set to false, the first app launched does not exit when the
ESCAPE or EXIT keys are pressed. Instead, the key is passed to the
app so it can choose what it wants to do with it. This is an easy
way to keep too many ESCAPE/EXIT key presses from quitting tp.