[rex-iphone] iOS app now defaults to current running app.
Previously as soon as a TV was selected it would list the apps of the available TV. Clicking on an app would launch it regardless of whether that app was already running, which it would then relaunch the app. Now if there is a current running app (besides the "Empty" app) the iOS controller app will go straight to the GestureViewController which displays the graphics called from the trickplay app and sends events to Trickplay.