[rex-iphone] Added changes that will make things more reliable and smoother.
Previously app would stay in App Browser screen even if the Trickplay app was closed and/or the socket connection to Trickplay was closed, now App Browser detects this change and pops back to the TV browser. Additionally, when a socket connection closes the whole SocketManager class is destroyed. Portions of the code that are socket reliant in the App Browser were previously checking for this classes existance before executing code, however, the class would not delete. This way if the connection closes the app should pop back to the TV browser rather than try to speak to a connection that doesn't exist.