[rex-iphone] Created the basic infrastructure for understanding the Advanced UI protocol and creating Trickplay-like objects specified by this protocol.
The "Advanced UI" protocol is the name given to the part of the Trickplay Controller protocol that allows creation and deletion of Trickplay-like objects (Rectangles, Images, Text, and Groups) on a supporting controller and getting/setting properties to these objects. The basic infrastructure is now established for understanding messages for this protocol. Currently the iOS "Take Control" app can create Rectangle objects of different size, position, and color on iOS via Trickplay controller Advanced UI protocol.