[rex-iphone] Added the ability to send images taken to Trickplay over protocol ver. 3.1. Additionally added the WM (welcome) and DG (drop resource group) features of protocol ver. 3.1.
The new protocol 3.1 added a welcome message WM which reveals an http port and version number to the controller. This message is received and the port number stored for use with interacting with the http service. Command DG was added which allows for deletion of cached resources which are associated with a group name as given in the new DR (declare resource) command in protocol 3.1. The controller now also responds to PI <path> command for Picking an image on the ipod controller. It should bring up the option to take an image or select an image from the iPod's library. Then the image is submitted to Trickplay via the path and the http port number received in the WM message. This is untested however.