[rex-iphone] Fixed the AudioController so that play_sound and stop_sound now function.
Previously controller:play_sound(string, int) and controller:stop_sound() on Trickplay were ignored because of a crash in the AudioController. The crash was found to be a bad access across the sound playing thread when the socket disconnects. This has been fixed and the AudioController no longer ignores play_sound and stop_sound (PS/SS).


Former-commit-id: 5efe06e559d7901ab7fb6927e7b53290a4771f92
Former-commit-id: 1a64a77889757a1909d5391befbd97329c903016