[rex-iphone] Fixed the AudioController so that play_sound and stop_sound now function.
Previously controller:play_sound(string, int) and controller:stop_sound() on Trickplay were ignored because of a crash in the AudioController. The crash was found to be a bad access across the sound playing thread when the socket disconnects. This has been fixed and the AudioController no longer ignores play_sound and stop_sound (PS/SS).