[rex-iphone] Fixed auto-resizing issues with AsyncImageView's; Added the ability to SET values for AdvancedUI Objects; Added the ability to CREATE/DESTROY/SET AdvancedUI Image Objects. BUGZID:947 BUGZID:949
Previously AsyncImageView Objects were resizing themselves incorrectly if being loaded for the first time (hence over the socket and not from a cache), this is fixed. TakeControl app now responds to the UX SET <json> command over the protocol allowing to set properties of AdvancedUI Objects (Trickplay elements). If the class of an AdvancedUI Object is "Image" TakeControl can now perform operations on this image (so far CREATE/DESTROY/SET function, GET is not implemented).