[rex-iphone] Full camera implementation finished. BUGZID:945 BUGZID:1019
Via the function controller:submit_picture({number w, number h}, bool editable, string mask_resource) the developer from Trickplay may command the iOS Device to bring up the camera or photo library. From there the user may select and image/take a picture. If a mask is given then it will display over the image as the user rotates, scales, and or moves the image via gesture motions. After selecting the desired transformations the image is cropped to the viewport. Lastly the image is scaled to the width and height specified by w and h. This transformed image is then sent back to Trickplay for for further use.