[engine] Added support for a list of attributes in app files.
The attributes can be read by the app itself using app.attributes and
are also available in lists of app info. Attributes are stored in the
system database after they are read from the app file. The first attribute
we are using is 'nolauncher' which tells the launcher to skip the app.
This is used for the empty app, the visual editor and the Burberry app.


Former-commit-id: 91ee10ead3a437ca3a1a5002c6502e1a1d55bf08
Former-commit-id: b79bac2119a6b03e4941a37ae805cecf40d64c94