[rex-iphone] A whole plethora of improvements were made.
Most of the improvements were simply to make the code easier to read and more consistant.
Applied method dispatching for AdvancedUI socket commands
Implemented the delete method. (only for clip for now, not sure if its needed elsewhere)
Added some extra properties from UIElement to AdvancedUI TrickplayUIElement Interface.
Made adjustments to how images are loaded to accomidate for tiling and in the future dynamic adjusting of the src image.
Tiling now is implemented but tiles vertically and horizontally symultaneously; cannot yet choose to tile just vertically or just horizontally.
Many other changes i definitely forgot but helped the integrity of the app.