[poker-iphone] Added some new phone assets. Fixed camera functionality image cropping and added new mask to camera editor on iphone. BUGZID:1018
Player could take pictures from their iphone to use as avatars in poker dawgz, but the cropping was screwed up and not proportional to the area on the TV screen that the image should have fit to. This is fixed. Additionally, a fancy new mask is sent to the iphone to assist the user in his/her scaling/cropping of their desire avatar image. This will allow the user to more easily position their desired avatar in a way which looks nice on their TV.