[rex-iphone] The necessary changes were made to convert the controller screen to a AdvancedUI Group object, thus allowing for access and manipulation emulating that of the global 'screen' property from Trickplay. BUGZID:1234
An AdvancedUI Group object now replaces the generic UIView used to place Advanced UI objects on the ios screen. Thus a developer can call methods such as do_add: or do_remove: remotely on the screen of the ios controller. This combined with the extensions to the Lua Controller API now gives the developer options to manipulate the ios controller's screen via <controller>.screen.<property/function call>. This access to the ios controller's screen as a AdvancedUI Group emulates the access to the global 'screen' as a Group on the TV (aka Trickplay).