[rex-iphone] Fix prevents controller from connecting to mDNS services that don't actually exist.
Found a bug where if Trickplay does not shut down its mDNS broadcast correctly the NetServiceBrowser may still pickup broadcast packets and assume a connection can be made. Though its not possible to force the controller to not print this service (since it can only assume that since it's receiving these packets a connection can be established), the controller takes a few extra steps to prevent from falling into an indeterminate state if the user selects to open that connection and instead attempts to reload active services.


Former-commit-id: bcca7bd23f1dcae69d850905d55f4bc3d7b65599
Former-commit-id: 579e97bf200e3d8633605f506facbc266fe9c143