[engine] Added support for modifier keys in events. bugzid:1124.
This change adds support for modifier keys in several events. It affects
UI element events, the public controller interface and controller events.

The following UIElement events now get an additional parameter:

	on_key_down
	on_key_up
	on_button_down
	on_button_up
	on_motion
	on_scroll

The additional parameter is nil if there are no modifiers down. Otherwise,
it is a table that can have any or all of the following keys set to true:

	shift,lock,control,super,meta,hyper

So, to check for the shift key down in one of these events, do this:

	function screen:on_button_down( x , y , button , clicks , modifier )
		if modifier and modifier.shift then
			...whatever...
		end
	end

Also, the following controller events get the same parameter:

	on_key_down
	on_key_up
	on_accelerometer
	on_pointer_move
	on_pointer_button_down
	on_pointer_button_up
	on_touch_down
	on_touch_move
	on_touch_up

Note that in our current implementation for iOS, these don't make sense, so
the modifier parameter will be nil.

The following public OEM interface functions have changed, which will cause
breakage with just about everybody (although very minor).

	tp_controller_key_down
	tp_controller_key_up
	tp_controller_accelerometer
	tp_controller_pointer_move
	tp_controller_pointer_button_down
	tp_controller_pointer_button_up
	tp_controller_touch_down
	tp_controller_touch_move
	tp_controller_touch_up


Former-commit-id: fcd3b5f36dce5aef07c60f7425021f303bf8d639
Former-commit-id: 30b95ab8501be59288694ea2f23eabe61ce5c5e8