[engine] Added GL texture cache.
If the cache is enabled, images loaded from files are cached automatically. The
cache is system-wide, so when you launch an app a second time and some of its
images are cached, you get a startup speed boost.

There is also a global called ImageCache which lets you explicitly cache images.
This is useful for images created from Bitmap or Canvas, which don't have a
natural cache key.

Setting TP_TEXTURE_CACHE_ENABLED to true, turns on the behavior. You still have
to set TP_TEXTURE_CACHE_LIMIT to the number of MB of textures to keep in the cache.
The cache is disabled by default and the default limit is 0 (which is the same as
disabling the cache).