[engine] Added mDNS service for _trickplay-http._tcp. UPnP discovery now advertises the HTTP port. Added /controllers HTTP/JSON API that returns the controllers port and protocol version. bugzid: 2172
This change implements the engine side of switching to advertising the HTTP port.
It does so in a backward-compatible way by adding a new mDNS service for HTTP, leaving
the existing service (_tp-remote._tcp) as it was. There are no changes to the
controller protocol.

Now, a client can discover a TrickPlay instance, and use its HTTP services without dealing
with the controller protocol at all.