[engine] Debugger can handle some commands when it is not in a break.
For example, by handling the "bn" command while not in a break, you can
cause the debugger to break. You can also edit breakpoints when not in
a break - which could cause the debugger to break.

The following commands are affected:

r, bn, q, b, a, bb, d and f.


Former-commit-id: 33d4672e917692d5b0779d192bf7e450a63987e3
Former-commit-id: d2cf98531930086b0103fec8e453e09284486ee9