[engine] Debugger can handle some commands when it is not in a break.
For example, by handling the "bn" command while not in a break, you can
cause the debugger to break. You can also edit breakpoints when not in
a break - which could cause the debugger to break.

The following commands are affected:

r, bn, q, b, a, bb, d and f.