[debugger] fixed default remote device setting bug. bugzid : 2861 Remote device drop down should always default to Emulator when current device disappears. 1) Run a separate instance of Trickplay on your PC with controllers_enabled=true. 2) Verify there is at least one more remote devices showing. 3) Select your local instance in the list. 4) Exit out of your local instance of TP.
Result: The list of remote devices does not always to back to the Emulator. I think it goes to the one that was above the one that was just exited.

Expected Result: It should always go back to Emulator. The developer may not realize and push an app to another remote device unexpectedly.


Former-commit-id: 33ba8a96b0e39c50c8208b2bc24122992216bf3f
Former-commit-id: 949aadcb09fa2e481e78f8d42d35658cc47ea83d