[engine] App tracks global values added by the user's code.
We install a __newindex function on the globals table that catches
any global variables added by the user. We capture the name of the
global as well as the line and file where it was added.

This does not catch cases where the user's code replaces an existing
global function with his own. For example:

local _my_clone = Clone
Clone = function(...) _my_clone(...) end

The new "Clone" function is not caught by this scheme.