[engine] Cherry pick engine change to editor.lb to allow finding the native path of the current app
[qa] New unit-test harness and framework
The goal is to improve automate unit testing.  The current harness can fail tests if the system is overloaded.  Timing for each test is arbitrary.  This new harness asks the tests to let us know when they're done.  We can still set a timeout to determine "stuck" tests if necessary, but using this method means that test which merely run long won't get failed for not having completed within some arbitrary time window when they are testing something asynchronous.
All tests will need to be ported to this new structure, but it shouldn't be hard and should result in major improvements to test reliability as well as easier construction/addition of new tests.